code/server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
requests
firebase-admin
cachetools
pydantic
python-multipart
motor
pymongo
